+++ libsplinter/src/network/peer.rs
@@ -11,6 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+// This module is deprecated
+
 use std::fmt;
 use std::sync::{Arc, Mutex};
