+++ libsplinter/src/network/sender.rs
@@ -11,6 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+// This module is deprecated
+
 use std::sync::mpsc::{channel, Sender};
 
 use crate::network::Network;
